Added mapping support for env vars
This commit allows environment variables to be used to set config
entries. This is done by adding a `env_var` field to
`Conform.Schema.Mapping` and logic to `Conform.Translate` to use env var
values when set.

If a config entry has an environment variable mapping then the value of
the environment variable, if it's set, will override all other values.